refactor: move showAsJson to separate file
